refactor load store register unsigned immedate, pre- post-indexed instructions
