+++ src/types.rs
@@ -32,6 +32,9 @@ pub type Imm5 = u8;
 pub type UImm64 = u64;
 pub type UImm32 = u32;
 pub type UImm16 = u16;
+pub type UImm15 = u16;
+pub type UImm14 = u16;
+pub type UImm13 = u16;
 pub type UImm12 = u16;
 pub type UImm10 = u16;
 pub type UImm6 = u8;
